[test-suite] Fix adding the halide runtime file as cxx after file renaming.
Summary:
In order for the bitcode files to be linked they need to be marked as being cxx language.
A previous patch renaming the halide runtime failed to do the renaming in the cmake file.
This should also fix the problem of failing aarch64 tests on green dragon - to be re-enabled after this fix.

Reviewers: MatzeB

Subscribers: aemerson, mehdi_amini, llvm-commits

Differential Revision: https://reviews.llvm.org/D23275